Reorganize classes into packages: providers, converters, and support
