+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/converter/ConvertWith.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package org.junit.jupiter.params;
+package org.junit.jupiter.params.converter;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
