+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/converter/package-info.java
@@ -1,6 +1,6 @@
 /**
- * {@link org.junit.jupiter.params.ArgumentConverter} implementations and their
- * corresponding {@link org.junit.jupiter.params.ConvertWith} annotations.
+ * {@link org.junit.jupiter.params.converter.ArgumentConverter} implementations and their
+ * corresponding {@link org.junit.jupiter.params.converter.ConvertWith} annotations.
  */
 
 package org.junit.jupiter.params.converter;
