+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/CsvArgumentsProvider.java
@@ -17,10 +17,7 @@ import com.univocity.parsers.csv.CsvParser;
 import com.univocity.parsers.csv.CsvParserSettings;
 
 import org.junit.jupiter.api.extension.ContainerExtensionContext;
-import org.junit.jupiter.params.AnnotationInitialized;
-import org.junit.jupiter.params.Arguments;
-import org.junit.jupiter.params.ArgumentsProvider;
-import org.junit.jupiter.params.support.ObjectArrayArguments;
+import org.junit.jupiter.params.support.AnnotationInitialized;
 
 class CsvArgumentsProvider implements ArgumentsProvider, AnnotationInitialized<CsvSource> {
