+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/EnumArgumentsProvider.java
@@ -17,10 +17,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import org.junit.jupiter.api.extension.ContainerExtensionContext;
-import org.junit.jupiter.params.AnnotationInitialized;
-import org.junit.jupiter.params.Arguments;
-import org.junit.jupiter.params.ArgumentsProvider;
-import org.junit.jupiter.params.support.ObjectArrayArguments;
+import org.junit.jupiter.params.support.AnnotationInitialized;
 
 class EnumArgumentsProvider implements ArgumentsProvider, AnnotationInitialized<EnumSource> {
