+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/MethodArgumentsProvider.java
@@ -20,10 +20,7 @@ import java.util.Iterator;
 import java.util.stream.Stream;
 
 import org.junit.jupiter.api.extension.ContainerExtensionContext;
-import org.junit.jupiter.params.AnnotationInitialized;
-import org.junit.jupiter.params.Arguments;
-import org.junit.jupiter.params.ArgumentsProvider;
-import org.junit.jupiter.params.support.ObjectArrayArguments;
+import org.junit.jupiter.params.support.AnnotationInitialized;
 import org.junit.platform.commons.JUnitException;
 import org.junit.platform.commons.util.PreconditionViolationException;
 import org.junit.platform.commons.util.ReflectionUtils;
