+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/ValueArgumentsProvider.java
@@ -19,10 +19,7 @@ import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
 import org.junit.jupiter.api.extension.ContainerExtensionContext;
-import org.junit.jupiter.params.AnnotationInitialized;
-import org.junit.jupiter.params.Arguments;
-import org.junit.jupiter.params.ArgumentsProvider;
-import org.junit.jupiter.params.support.ObjectArrayArguments;
+import org.junit.jupiter.params.support.AnnotationInitialized;
 import org.junit.platform.commons.util.Preconditions;
 
 class ValueArgumentsProvider implements ArgumentsProvider, AnnotationInitialized<ValueSource> {
