+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/package-info.java
@@ -1,6 +1,6 @@
 /**
- * {@link org.junit.jupiter.params.ArgumentsProvider} implementations and their
- * corresponding {@link org.junit.jupiter.params.ArgumentsSource} annotations.
+ * {@link org.junit.jupiter.params.provider.ArgumentsProvider} implementations and their
+ * corresponding {@link org.junit.jupiter.params.provider.ArgumentsSource} annotations.
  */
 
 package org.junit.jupiter.params.provider;
