+++ junit-jupiter-params/src/test/java/org/junit/jupiter/params/provider/CsvArgumentsProviderTests.java
@@ -17,7 +17,6 @@ import static org.mockito.Mockito.when;
 import java.util.stream.Stream;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.Arguments;
 
 class CsvArgumentsProviderTests {
